+++ tests/functions.php
@@ -87,6 +87,11 @@ function generate_files2(): void
     @chmod($files2Path . '/dir6/exec.txt', 0755);
 }
 
+function generate_files3(): void
+{
+    file_put_contents(TEST_ROOT . '/data/empty_file.txt', '');
+}
+
 function get_words(): array
 {
     $wordsFile = TEST_ROOT . '/data/words.txt';
@@ -123,7 +128,7 @@ function export_test_data(mixed $data): string
         return $export;
     }
 
-    if (\is_string($data) && !ctype_print($data)) {
+    if (\is_string($data) && $data !== '' && !ctype_print($data)) {
         return 'base64_decode("' . base64_encode($data) . '")';
     }